requirements.txt
psycopg2
psycopg2-binary
djangorestframework
pillow
